+++ eventkit_cloud/ui/static/ui/app/tests/StatusDownloadPage/DataCartDetails.spec.tsx
@@ -8,6 +8,9 @@ import DataPackOptions from '../../components/StatusDownloadPage/DataPackOptions
 import DataPackGeneralTable from '../../components/StatusDownloadPage/DataPackGeneralTable';
 import DataCartInfoTable from '../../components/StatusDownloadPage/DataCartInfoTable';
 import { DataCartDetails } from '../../components/StatusDownloadPage/DataCartDetails';
+import * as utils from "../../utils/generic";
+import {shallow} from "enzyme";
+import {ExportAOI} from "../../components/CreateDataPack/ExportAOI";
 
 describe('DataCartDetails component', () => {
     let shallow;
@@ -34,6 +37,7 @@ describe('DataCartDetails component', () => {
                 groups: {},
                 members: {},
             },
+            extent: { type: 'FeatureCollection', features: [] }
         },
         provider_tasks: [
             {
@@ -105,8 +109,11 @@ describe('DataCartDetails component', () => {
     let wrapper;
     let instance;
     const setup = (overrides = {}) => {
+         const config = {
+            BASEMAP_URL: 'http://my-osm-tile-service/{z}/{x}/{y}.png',
+        };
         props = { ...getProps(), ...overrides };
-        wrapper = shallow(<DataCartDetails {...props} />);
+        wrapper = shallow(<DataCartDetails {...props} />, {context: { config },});
         instance = wrapper.instance();
     };
